[SERVICE-292] Move scope schema into layout-config
