+++ res/provider/config/layouts-config.schema.json
@@ -119,6 +119,9 @@
                 },
                 {
                     "$ref": "#/definitions/rule-other"
+                },
+                {
+                    "$ref": "#/definitions/scope"
                 }
             ]
         },
@@ -201,6 +204,82 @@
                 {
                     "$ref": "#/definitions/configuration-object"
                 }
+            ],
+            "additionalProperties": false
+        },
+        "scope": {
+            "title": "Scope",
+            "type": "object",
+            "oneOf": [
+                {
+                    "$ref": "#/definitions/scope-app"
+                },
+                {
+                    "$ref": "#/definitions/scope-window"
+                },
+                {
+                    "$ref": "#/definitions/scope-other"
+                }
+            ]
+        },
+        "scope-app": {
+            "title": "ApplicationScope",
+            "description": "A scope targeting a specific OpenFin application and all of its windows",
+            "properties": {
+                "level": {
+                    "type": "string",
+                    "enum": [
+                        "application"
+                    ]
+                },
+                "uuid": {
+                    "type": "string"
+                }
+            },
+            "required": [
+                "level",
+                "uuid"
+            ],
+            "additionalProperties": false
+        },
+        "scope-window": {
+            "title": "WindowScope",
+            "description": "A scope targeting a specific OpenFin window",
+            "properties": {
+                "level": {
+                    "type": "string",
+                    "enum": [
+                        "window"
+                    ]
+                },
+                "uuid": {
+                    "type": "string"
+                },
+                "name": {
+                    "type": "string"
+                }
+            },
+            "required": [
+                "level",
+                "uuid",
+                "name"
+            ],
+            "additionalProperties": false
+        },
+        "scope-other": {
+            "title": "OtherScope",
+            "description": "A scope for all levels with no extra data required (e.g. desktop-level configuration)",
+            "properties": {
+                "level": {
+                    "type": "string",
+                    "enum": [
+                        "service",
+                        "desktop"
+                    ]
+                }
+            },
+            "required": [
+                "level"
             ],
             "additionalProperties": false
         }
