+++ src/provider/config/Watch.ts
@@ -1,5 +1,4 @@
-import {Rule} from '../../../gen/provider/config/layouts-config';
-import {Scope} from '../../../gen/provider/config/scope';
+import {Rule, Scope} from '../../../gen/provider/config/layouts-config';
 import {Signal2} from '../Signal';
 
 import {ConfigUtil, Mask} from './ConfigUtil';
