+++ src/provider/model/DesktopTabGroup.ts
@@ -1,6 +1,6 @@
 import {_Window} from 'hadouken-js-adapter/out/types/src/api/window/window';
 
-import {Scope} from '../../../gen/provider/config/scope';
+import {Scope} from '../../../gen/provider/config/layouts-config';
 
 import {ApplicationUIConfig, TabAddedPayload, TabGroupEventPayload, TabProperties, WindowIdentity, WindowState} from '../../client/types';
 import {WindowMessages} from '../APIMessages';
