+++ src/provider/model/DesktopTabstripFactory.ts
@@ -1,7 +1,6 @@
 import {_Window} from 'hadouken-js-adapter/out/types/src/api/window/window';
 
-import {ConfigurationObject, Tabstrip} from '../../../gen/provider/config/layouts-config';
-import {Scope} from '../../../gen/provider/config/scope';
+import {ConfigurationObject, Scope, Tabstrip} from '../../../gen/provider/config/layouts-config';
 import {ApplicationUIConfig} from '../../client/types';
 import {ScopedConfig} from '../config/Store';
 import {MaskWatch} from '../config/Watch';
